Reworked Verify and DCHeader
Reworked how 2nd tension is displayed in verify, should be more accurate and take into account the info stored in the DB (only the last test is saved in T1, but the change from the 2nd to last test to the last test is saved in T2, I thought originally that second to last test was saved in T1 and instead of saving the last test, only the change from second to last -> last was saved in T2.)
Also added 2ndTT.py to check how far away a tube's 2nd tension date is from just a scan. Tried to pickle data with no luck.